// src/modules/seller/financial-hub/src/interface/invoice-center-interface.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export interface IInvoiceCenterResponse {
  id: string;
  type: string;
  createdAt: string;
  fromName: string;
  fromAddress: string;
  fromMobile: string;
  fromEmail: string;
  toName: string;
  toAddress: string;
  toEmail: string;
  toMobile: string;
  subTotalAmount: number;
  discountAmount: number;
  totalAmount: number;
  currency: string;
  earning?: Earning;
  withdrawal?: Withdrawal;
  items: Item[];
}

interface Earning {
  id: string;
  status: string;
  purchasedAt: string;
  completedAt: string;
  releasedAt: null;
  currency: string;
  grossAmount: number;
  serviceFee: number;
  serviceFeeRate: number;
  paymentFeeRate: number;
  paymentFee: number;
  netAmount: number;
  order: Order;
}

interface Order {
  id: string;
  sellerEntityId: string;
}

export interface Item {
  id: string;
  no: number;
  description: string;
  priceRate: number;
  price: number;
  quantity: number;
  amount: number;
}

interface Withdrawal {
  id: string;
  status: string;
  createdAt: string;
  withdrawnAt: string;
  withdrawMethod: string;
  createdBy: string;
  grossAmount: number;
  feeAmount: number;
  netAmount: number;
  currency: string;
  paymentId: string;
  bankAccount: BankAccount;
}

interface BankAccount {
  id: string;
  entityId: string;
}